+++ src/main/java/io/github/ducomin/VendasApplication.java
@@ -1,62 +1,46 @@
 package io.github.ducomin;
 
-import java.util.List;
-
+import io.github.ducomin.domain.entity.Cliente;
+import io.github.ducomin.domain.entity.Pedido;
+import io.github.ducomin.domain.repository.Clientes;
+import io.github.ducomin.domain.repository.Pedidos;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
-import io.github.ducomin.domain.entity.Cliente;
-import io.github.ducomin.repository.Clientes;
+import java.math.BigDecimal;
+import java.time.LocalDate;
 
 @SpringBootApplication
 public class VendasApplication {
 
 	@Bean
-	public CommandLineRunner init(@Autowired Clientes clientes){
+	public CommandLineRunner init(
+			@Autowired Clientes clientes,
+			@Autowired Pedidos pedidos
+	) {
 		return args -> {
 			System.out.println("Salvando clientes");
-			clientes.save(new Cliente("Dougllas"));
-			clientes.save(new Cliente("Outro Cliente"));
+			Cliente fulano = new Cliente("Fulano");
+			clientes.save(fulano);
 
-			List<Cliente> todosClientes = clientes.findAll();
-			todosClientes.forEach(System.out::println);
+			Pedido p = new Pedido();
+			p.setCliente(fulano);
+			p.setDataPedido(LocalDate.now());
+			p.setTotal(BigDecimal.valueOf(100));
 
-			System.out.println("Atualizando clientes");
-			todosClientes.forEach(c -> {
-				c.setNome(c.getNome() + " atualizado.");
-				clientes.save(c);
-			});
+			pedidos.save(p);
 
-			todosClientes = clientes.findAll();
-			todosClientes.forEach(System.out::println);
+			//            Cliente cliente = clientes.findClienteFetchPedidos(fulano.getId());
+			//            System.out.println(cliente);
+			//            System.out.println(cliente.getPedidos());
 
-			System.out.println("Buscando clientes");
-			clientes.findByNomeLike("Cli").forEach(System.out::println);
+			pedidos.findByCliente(fulano).forEach(System.out::println);
 
-			System.out.println("Buscando clientes (query)");
-			clientes.encontrarPorNome("Cli").forEach(System.out::println);
 
-			todosClientes = clientes.findAll();
-			if(todosClientes.isEmpty()){
-				System.out.println("Nenhum cliente encontrado.");
-			}else{
-				todosClientes.forEach(System.out::println);
-			}
 
-			System.out.println("deletando clientes");
-			clientes.findAll().forEach(c -> {
-				clientes.delete(c);
-			});
-
-			todosClientes = clientes.findAll();
-			if(todosClientes.isEmpty()){
-				System.out.println("Nenhum cliente encontrado.");
-			}else{
-				todosClientes.forEach(System.out::println);
-			}
 		};
 	}
