+++ src/main/java/io/github/ducomin/domain/repository/ItemsPedido.java
@@ -1,0 +1,7 @@
+package io.github.ducomin.domain.repository;
+
+import io.github.ducomin.domain.entity.ItemPedido;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemsPedido extends JpaRepository<ItemPedido, Integer> {
+}
