+++ src/main/java/io/github/ducomin/domain/repository/Produtos.java
@@ -1,0 +1,8 @@
+package io.github.ducomin.domain.repository;
+
+import io.github.ducomin.domain.entity.Produto;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface Produtos extends JpaRepository<Produto,Integer> {
+}
