--- src/main/java/io/github/ducomin/repository/ItemsPedido.java
@@ -1,7 +1,0 @@
-package io.github.ducomin.repository;
-
-import io.github.ducomin.domain.entity.ItemPedido;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ItemsPedido extends JpaRepository<ItemPedido, Integer> {
-}
