--- src/main/java/io/github/ducomin/repository/Produtos.java
@@ -1,7 +1,0 @@
-package io.github.ducomin.repository;
-
-import io.github.ducomin.domain.entity.Produto;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface Produtos extends JpaRepository<Produto,Integer> {
-}
